Java Essential - hw2 - upd task 5

+++ src/javaEssential/homework/hw2/task5/MyArea.java
@@ -3,7 +3,7 @@ package javaEssential.homework.hw2.task5;
 public class MyArea {
     private static final float PI = 3.14f;
 
-    public static double areaOfCircle(float radius) {
+    public static double areaOfCircle(double radius) {
         return PI * Math.pow(radius,2);
     }
 }
